+++ src/Request.php
@@ -22,10 +22,4 @@ interface Request {
 	 */
 	public function getHeaders();
 
-	/**
-	 * @since 0.4
-	 * @return RequestOptions
-	 */
-	public function getOptions();
-
 }