+++ tests/unit/FluentRequestTest.php
@@ -12,7 +12,6 @@ class FluentRequestTest extends \PHPUnit_Framework_TestCase {
 
 		$this->assertEquals( array(), $request->getParams() );
 		$this->assertEquals( array(), $request->getHeaders() );
-		$this->assertInstanceOf( 'Mediawiki\Api\RequestOptions', $request->getOptions() );
 	}
 
 	public function testSetParams() {
@@ -31,25 +30,6 @@ class FluentRequestTest extends \PHPUnit_Framework_TestCase {
 		$request->setHeaders( $params );
 
 		$this->assertEquals( $params, $request->getHeaders() );
-	}
-
-	public function testSetOptions() {
-		$request = new FluentRequest();
-
-		$newOptions = new RequestOptions();
-		$this->assertNotSame( $newOptions, $request->getOptions() );
-
-		$request->setOptions( $newOptions );
-
-		$this->assertSame( $newOptions, $request->getOptions() );
-	}
-
-	public function testSetAttempts() {
-		$request = new FluentRequest();
-
-		$request->setAttempts( 12 );
-
-		$this->assertEquals( 12, $request->getOptions()->getAttempts() );
 	}
 
 }